only add files we were able to get data for

+++ generate.sh
@@ -12,5 +12,7 @@ OUTPUT_FILE_NAME=$(echo "$2" | sed 's/ /_/')
 
 ./github-top --token "$1" --preset "$2" --output yaml --file "$OUTPUT_FILE_NAME.yml"
 
-echo "page: $OUTPUT_FILE_NAME.html\ntitle: $3" | cat - "$OUTPUT_FILE_NAME.yml" > "_data/locations/$OUTPUT_FILE_NAME.yml"
-echo "---\ntype: location\nlocation: $OUTPUT_FILE_NAME\n---" > "$OUTPUT_FILE_NAME.md"
+if [ -f "$OUTPUT_FILE_NAME.yml" ]; then
+  echo "page: $OUTPUT_FILE_NAME.html\ntitle: $3" | cat - "$OUTPUT_FILE_NAME.yml" > "_data/locations/$OUTPUT_FILE_NAME.yml"
+  echo "---\ntype: location\nlocation: $OUTPUT_FILE_NAME\n---" > "$OUTPUT_FILE_NAME.md"
+fi
